requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0

# 腾讯云语音识别SDK
tencentcloud-sdk-python>=3.0.0

# WebSocket支持
websockets>=10.0
flask-socketio>=5.3.0

# 音频处理
pydub>=0.25.0
numpy>=1.21.0

# 额外的音频处理工具
librosa>=0.9.0 